# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample S⁴ server. Use a hashtag `#s4` in your e-tickes from SAC or SCCD.",
    "version": "1.0.0",
    "title": "S⁴ - Sugestão de Solução de Salts na Sustentação",
    "contact": {
      "email": "wilian.ivo@softplan.com.br"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "127.0.0.1:8081",
  "basePath": "/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/server/avaliar": {
      "post": {
        "tags": [
          "server"
        ],
        "summary": "Validate e-tickets",
        "description": "This end-point return result about avaliation of e-tickets",
        "operationId": "avaliar",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Number and item e-ticket ",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "atendimento": {
                  "type": "integer",
                  "format": "int64"
                },
                "item": {
                  "type": "string",
                  "format": "int64"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
...
}